requirements.txt
alpaca-trade-api==3.2.0
pandas==2.2.2
numpy==1.26.4
aiohttp==3.9.5
PyYAML==6.0.1
requests==2.32.2
python-dotenv==1.1.0
finvizfinance==0.14
matplotlib==3.8.2
plotly==5.17.0
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
PyJWT==2.8.0
openai==1.3.0
nltk==3.8.1
yfinance==0.2.18
scipy==1.11.4
passlib[bcrypt]==1.7.4
